sdl: remove -alt-grab and -ctrl-grab support
With absolute pointer devices such as usb-tablet being widely used
mouse grabs (for relative pointing devices) should be rarely needed
these days.  So the benefit of the options to configure the hotkey
modifiers for grab (and other actions) seems questionable.  Which
is expecially true for the -ctrl-grab which isn't handled in the
handle_keyup() code.

Drop the code, print a notice when the option is still used.

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>